mingus/stable-requirements.txt
BeautifulSoup==3.1.0.1
http://dist.repoze.org/PIL-1.1.6.tar.gz
python-dateutil==1.4.1
Django
django-debug-toolbar==0.8.0
django-tagging==0.3
Pygments==1.0
feedparser==4.1
Markdown==2.0.1
textile==2.1.3
django-markup==0.3
http://www.aeracode.org/releases/south/south-0.6.1.tar.gz

-e git://github.com/django-extensions/django-extensions.git@664ca825578c5c13f0817eba0578fbda4bac62cf#egg=django_extensions
-e svn+http://django-template-utils.googlecode.com/svn/trunk@109#egg=template_utils-0.4p2-py2.5-dev
-e hg+https://bitbucket.org/pmarti/django-proxy/@tip#egg=django_proxy-tip
-e git://github.com/montylounge/django-google-analytics.git@6b7aa6dba851dd59dca0a20bc7eefe70754384e4#egg=google_analytics
-e git://github.com/montylounge/django-basic-apps.git@e39b6cd1400176f3635f4bed53ebc97724ef2a3b#egg=django_basic_apps

-e git://github.com/mintchaos/django_compressor.git@d5d28eb57323a7e88356071c8c4af5792d2b837d#egg=django_compressor
-e hg+https://bitbucket.org/ubernostrum/django-contact-form/@tip#egg=django_contact_form-tip
-e svn+http://django-db-log.googlecode.com/svn/trunk@13#egg=django_db_log-1.0.1-py2.5-dev
-e git://github.com/montylounge/django-disqus.git@35a022f28c50b9f7cdfcdd5deb4ae6f9150f4358#egg=django_disqus
-e git://github.com/zerok/django-flatblocks.git@588d49c69e290ba7429559c64105102f88fb6fcd#egg=django_flatblocks
-e git://github.com/sunlightlabs/django-honeypot.git@825c1cd0fc99491d5f30b976f511a97b56ca9978#egg=django_honeypot8
-e svn+http://django-navbar.googlecode.com/svn/trunk@13#egg=django_navbar-0.1.0-py2.5-dev
-e git://github.com/brosner/django-oembed.git@0b09321a6042c490534e5940083136f712da57b0#egg=django_oembed
-e hg+https://bitbucket.org/jezdez/django-robots/@tip#egg=django_robots-tip
-e git://github.com/pmarti/django-sugar.git@32a522d8ea7144414e41cf2d151692e3ab68dac7#egg=django_sugar
-e hg+https://bitbucket.org/kmike/django-view-cache-utils/@tip#egg=django_view_cache_utils-tip
-e hg+https://sorl-thumbnail.googlecode.com/hg/@tip#egg=sorl_thumbnail-tip
